% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{centreLawnOnPosition}
\alias{centreLawnOnPosition}
\title{Centre a regular rectangular survey on a given position}
\usage{
centreLawnOnPosition(centreLon, centreLat, proj4string = NULL,
  tolerance = 20, ...)
}
\arguments{
\item{centreLon}{Desired centre location of survey}

\item{centreLat}{Desired centre location of survey}

\item{proj4string}{projection string of class \link{CRS-class}.  If NULL defaults to "+proj=longlat +datum=WGS84"}

\item{tolerance}{maximum distance (in metres) between desired survey centre and realised survey centre}

\item{...}{other arguments to be passed into \link{lawnSurvey}}
}
\value{
Line transect coordinates (lon, lat) as specified in \link{lawnSurvey}
}
\description{
Centres a regular rectangular survey on a desired latitude and longitude
}
\examples{
\dontrun{
coords=centreLawnOnPosition(centreLon=-170,centreLat=-60,lineLengthkm=2,lineSpacingkm=0.5,
startBearingdeg=30,numOfLines=5)
plot(0,0,xlim=range(coords[,1]),ylim=range(coords[,2]),type='n',
xlab='Longitude, deg',ylab='Latitude, deg')
arrows(x0=coords[1:(nrow(coords)-1),1], y0=coords[1:(nrow(coords)-1),2], 
      x1 = coords[2:nrow(coords),1], y1 = coords[2:nrow(coords),2])
text(coords,row.names(coords),cex=0.6)
points(coords[1,1],coords[1,2],col='blue',pch=17,cex=2)
points(coords[nrow(coords),1],coords[nrow(coords),2],col='blue',pch=15,cex=2)
points(-170,-60,col='purple',pch=19,cex=2)
points(geomean(coords),col='red',pch=19,cex=1)
legend('bottomright',c('Beginning','End','Desired centre','Actual centre'),
 col=c('blue','blue','purple','red'),pch=c(17,15,19,19),pt.cex=c(1,1,2,1))
}
}
