% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineEVL.R
\name{combineEVL}
\alias{combineEVL}
\title{Combine Echoview line objects}
\usage{
combineEVL(evlL)
}
\arguments{
\item{evlL}{list of objects of class evl, typically resulting from a class of readevl2R}
}
\value{
an evl class object
}
\description{
This function combines evl class objects into a single .evl file
}
\seealso{
\code{\link{readevl2R}} \code{\link{writeRevl2file}}
}
