% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVAcoVarCheck.R
\name{EVAcoVarCheck}
\alias{EVAcoVarCheck}
\title{Check an Echoview acoustic variable}
\usage{
EVAcoVarCheck(EVFile, acoVarName)
}
\arguments{
\item{EVFile}{(COM object) of the open Echoview File}

\item{acoVarName}{(character) Acoustic variable name}
}
\value{
(list) $useable (boolean) $measurementCount (integer) number of pings in the acoustic variable NA if unuseable $msg (character) vector of messages
}
\description{
This function checks if an Echoview acoustic variable is useable and returns the number of pings in the variable.
}
