% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVAcoVarConstantEdit.R
\name{EVAcoVarDataGenerator}
\alias{EVAcoVarDataGenerator}
\title{Control an Echoview data generator variable using COM scripting.}
\usage{
EVAcoVarDataGenerator(EVFile, acoVarName, Algorithm = c("Constant",
  "Tvg", "ConstantTvg"), OutputType = c("Sv", "TS", "Other dB",
  "Linear"), constantValue = NULL, ApplyTvgOverride = NULL,
  TvgOverride = NULL)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{acoVarName}{Acoustic variable name}

\item{Algorithm}{=c("Constant","Tvg","ConstantTvg") Data generator algorithm type}

\item{OutputType}{=c("Sv","TS","Other dB","Linear") output type}

\item{constantValue}{= NULL (numerical) constant value for the data generator.}

\item{ApplyTvgOverride}{= NULL (boolean) Appply TVG override}

\item{TvgOverride}{= NULL (numerical) TVG override value.}
}
\value{
a list object with two elements.  $success: ping subset successfully changed, and $msg: message for processing log.
}
\description{
This function modifies the data generator settings in an existing Echoview data generator virtual variable.
}
\examples{
\dontrun{
EVAcoVarDataGenerator(EVFile=EVFile,acoVarName='Data generator 1',
Algorithm='Constant',
OutputType='Linear',
constantValue=0.5,
ApplyTvgOverride=NULL,
TvgOverride=NULL)

}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
