% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVAcoVarNameFinder}
\alias{EVAcoVarNameFinder}
\title{Find an acoustic variable by name}
\usage{
EVAcoVarNameFinder(EVFile, acoVarName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{acoVarName}{The name of an acoustic variable in the Echoview file}
}
\value{
a list object with two elements. $EVVar: An Echoview acoustic variable object, and $msg: message for processing log.
}
\description{
This function finds an acoustic variable in an Echoview file by name and 
returns the variable pointer.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
varObj <- EVAcoVarNameFinder(EVFile, "120 7x7 convolution")$EVVar
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
