% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVAddCalibrationFile}
\alias{EVAddCalibrationFile}
\title{Add a calibration file (.ecs) to a fileset}
\usage{
EVAddCalibrationFile(EVFile, filesetName, calibrationFile)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{filesetName}{An Echoview fileset name}

\item{calibrationFile}{An Echoview calibration (.ecs) file path and name}
}
\value{
a list object with one element. $msg: message for processing log
}
\description{
This function adds a calibration file (.ecs) to a fileset using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVAddCalibrationFile(EVFile = EVFile, filesetName = '038-120-200', 
calibrationFile = '~~/KAOS/20120326_KAOS_SimradEK5.ecs')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
