% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVAddRawData}
\alias{EVAddRawData}
\title{Add raw data files to an open Echoview file (.EV)}
\usage{
EVAddRawData(EVFile, filesetName, dataFiles)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{filesetName}{Echoview fileset name}

\item{dataFiles}{vector of full path and name for each data file.}
}
\value{
a list object with two elements.  $nbrFilesInFileset: number of raw data files in the fileset, and $msg: message for processing log.
}
\description{
This function adds raw data files to an open Echoview file (.EV) via COM scripting.  The function assumes the Echoview fileset name already exists.
}
\examples{
\dontrun{
filenamesV <- c('~~/KAOS/raw/L0055-D20030115-T171028-EK60.raw', 
                 '~~/KAOS/raw/L0055-D20030115-T182914-EK60.raw')
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVNewFile(EVAppObj,templateFn="~~/KAOS/KAOStemplate.EV")$EVFile
EVAddRawData(EVFile = EVFile, filesetName = '038-120-200', dataFiles = filenamesV)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVNewFile}}  \code{\link{EVCreateNew}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
