% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVAddVariable.R
\name{EVAddVariable}
\alias{EVAddVariable}
\title{Add a variable to an Echoview file}
\usage{
EVAddVariable(EVFile, acoVarName, EOperator, addedVarName = NULL)
}
\arguments{
\item{EVFile}{Echoview file COM object}

\item{acoVarName}{The acoustic variable name to which a new variable will be added.}

\item{EOperator}{The 'EOperator' number of the variable to add. See notes}

\item{addedVarName}{=NULL. Name of the added variable.}
}
\value{
list with three elements: $varObj acoustic variable COM object; $objName new name of the variable object;$msg message vector
}
\description{
This function adds an Echoview variable to an existing Echoview file.  The variable is added by specifying the existing variable to which the new variable is added and the new variable type.
}
\details{
The new variable type is specifed using the variable-specific 'EOperator' integer.   For example, the 'Ping subset' operator 'EOperator' is 62. See Echoview help for details.
}
\note{
This function is under development.  We hope to add an 'EOperator' look up table sp that the type of variable to add can be specified by its character name, rather than 'EOperator'
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\link{EVAcoVarNameFinder}
}
\author{
Martin Cox
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
