% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVAdjustDataRngBitmap}
\alias{EVAdjustDataRngBitmap}
\title{Change the data range bitmap of an acoustic object}
\usage{
EVAdjustDataRngBitmap(varObj, minRng, maxRng)
}
\arguments{
\item{varObj}{An Echoview acoustic variable COM object, perhaps resulting from a call of EVAcoVarNameFinder()}

\item{minRng}{the minimum data value to set}

\item{maxRng}{the maximum data value to set}
}
\value{
a vector of pre- and post-function call data range settings
}
\description{
This function changes the data range in an Echoview data range bitmap virtual variable
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
varObj <- EVAcoVarNameFinder(EVFile, acoVarName = "38 data range bitmap")$EVVar
EVAdjustDataRngBitmap(varObj, minRng = -90, maxRng = 0)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVAcoVarNameFinder}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
