% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVBottomDetection.R
\name{EVBottomDetection}
\alias{EVBottomDetection}
\title{Bottom Detection}
\usage{
EVBottomDetection(EVFile, EVVar, LineName = "Bottom", algorithm = NULL,
  StartDepth = NULL, StopDepth = NULL, MinSv = NULL,
  UseBackstep = NULL, DiscriminationLevel = NULL,
  BackstepRange = NULL, PeakThreshold = NULL, MaxDropouts = NULL,
  WindowRadius = NULL, MinPeakAssymmetry = NULL,
  replaceOldBottom = TRUE, SpanGaps = TRUE)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{EVVar}{An Echoview Variable, accepts inputs as Character, list or Variable object (COMIDispatch)}

\item{LineName}{Character of the output name for the detected Line}

\item{algorithm}{numeric [0 - 2] Defines which bottom detection algorithm should be used: 0 for Delta Sv, 1 for Maximum Sv, 2 for Best bottom Candidate}

\item{StartDepth}{numeric [m] Minimum bottom detection depth}

\item{StopDepth}{numeric [m] maximum bottom detection detpth}

\item{MinSv}{numeric [dB] minimum detection Sv}

\item{UseBackstep}{Boolean [True or False]}

\item{DiscriminationLevel}{numeric [dB] Minimum discrimination threshold}

\item{BackstepRange}{numeric [m] Backstep range}

\item{PeakThreshold}{numeric [dB] Threshold for peak detection}

\item{MinPeakAssymmetry}{numeric}

\item{replaceOldBottom}{Boolean (TRUE or FALSE) If TRUE and a line with the same name as LineName already exists, the old line will be overwritten with the new one}

\item{SpanGaps}{Boolean (TRUE or FALSE) Decides wheteher span gaps is activated or notm default=TRUE}

\item{MaxDropOuts}{numeric [samples] Maximum number of dropout samples before bottom detection fails}

\item{windowRadius}{numeric [samples] Search window size}
}
\description{
This function runs a bottom detection on an echogram
Bottom algorithms and settings are explained in the [Echoview Help File]('https://support.echoview.com/WebHelp/Reference/Algorithms/Line_picking_algorithm.htm')
}
\examples{
#To be added - Needs Example data
#Starting Echoview
echoview = StartEchoview()
#Create a new EV File
EVFile <- EVCreateNew(EVAppObj=echoview, dataFiles = "rawfile.raw")$EVFile
Varname <- "Fileset 1: Sv pulse compressed wideband pings T1"
bottom <- EVBottomDetection(EVFile, EVVar=Varname, LineName="Bottom")
# Change the algorithm to Best bottom candidate
bottom <- EVBottomDetection(EVFile, EVVar=Varname, LineName="Bottom",algorithm=2)

}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
