% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVBackgroundNoiseRemoval.R,
%   R/EVChangeVariableGrid.R
\name{EVChangeVariableGrid}
\alias{EVChangeVariableGrid}
\title{Adjust the Echoview virtual variable 'Background noise removal' using COM}
\usage{
EVChangeVariableGrid(
  EVFile,
  acoVarName,
  timeDistanceGridType = NULL,
  depthGridType = NULL,
  timeDistanceGridDistance = NULL,
  depthGridDistance = NULL,
  EVLineName = NULL,
  showenum = FALSE
)

EVChangeVariableGrid(
  EVFile,
  acoVarName,
  timeDistanceGridType = NULL,
  depthGridType = NULL,
  timeDistanceGridDistance = NULL,
  depthGridDistance = NULL,
  EVLineName = NULL,
  showenum = FALSE
)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{acoVarName}{(character) Echoview acoustic variable name}

\item{timeDistanceGridType}{(numeric) specifying the along track grid type. 0 = no grid, 1 = time (minutes), 2 = GPS distance (NMi), 3 = Vessel Log Distance (NMi), 4 = Pings, 5 = GPS distance (m), 6 = Vessel Log Distance (m).}

\item{depthGridType}{(numeric) 0 = no grid, 1 = depth grid, 2 = use reference line.}

\item{timeDistanceGridDistance}{(numeric) vertical grid line spacing. Not needed if verticalType = 0.}

\item{depthGridDistance}{(numeric) horizontal grid line spacing. Ignored if horizontalType = 0.}

\item{EVLineName}{(character) an Echoview line object name. Not needed if depthGridType != 2.}

\item{showenum}{= FALSE (boolean) if TRUE Echoview grid enum values are returned to the workspace.  All other arguments are ignored.}
}
\value{
a list object with three elements.  $predataRangeSettings: a vector of pre-function call settings; $postDataRangeSettings a vector of post-function call settings, and $msg: message for processing log.

a list object with three elements.  $predataRangeSettings: a vector of pre-function call settings; $postDataRangeSettings a vector of post-function call settings, and $msg: message for processing log.
}
\description{
This function adjusts the settings of a Background Noise Removal virtual variable using COM scripting.

This function sets the grid separation and depth reference line for an acoustic variable using COM scripting.
}
\details{
\code{depthGridType} is specified in the Echoview COM enum values called 'EDepthRangeGridMode'. \code{TimeDistanceGridMode} is specified in the Echoview COM enum values called ETimeDistanceGridMode

\code{depthGridType} is specified in the Echoview COM enum values called 'EDepthRangeGridMode'. \code{TimeDistanceGridMode} is specified in the Echoview COM enum values called ETimeDistanceGridMode
}
\note{
This function was completely rewritten Sept. 2017 which resulted in changes to function argument names.

This function was completely rewritten Sept. 2017 which resulted in changes to function argument names.
}
\examples{
\dontrun{
EVAppObj <- start
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

EVChangeVariableGrid <- function (EVFile, 
acoVarName="38 seabed and surface excluded", 
timeDistanceGridType=5, depthGridType=2, 
timeDistanceGridDistance = 100, depthGridDistance = 10, 
EVLineName = "Fixed depth 250 m")
#remove grid:
EVChangeVariableGrid <- function (EVFile, 
acoVarName="38 seabed and surface excluded", 
timeDistanceGridType=0, depthGridType=0)
#only show Echoview grid enum values
EVChangeVariableGrid(showenum = TRUE)
}
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

EVChangeVariableGrid <- function (EVFile, 
acoVarName="38 seabed and surface excluded", 
timeDistanceGridType=5, depthGridType=2, 
timeDistanceGridDistance = 100, depthGridDistance = 10, 
EVLineName = "Fixed depth 250 m")
#remove grid:
EVChangeVariableGrid <- function (EVFile, 
acoVarName="38 seabed and surface excluded", 
timeDistanceGridType=0, depthGridType=0)
#only show Echoview grid enum values
EVChangeVariableGrid(showenum = TRUE)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}

\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVAcoVarNameFinder}} \code{\link{EVFindLineByName}}

\code{\link{EVAcoVarNameFinder}} \code{\link{EVFindLineByName}}
}
\author{
Martin Cox

Martin Cox
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
