% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVClearRawData}
\alias{EVClearRawData}
\title{Clear all files from a fileset}
\usage{
EVClearRawData(EVFile, filesetName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{filesetName}{An Echoview fileset name}
}
\value{
A list object with one element. $msg message for processing log
}
\description{
This function clears all .raw files from a fileset using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVClearRawData(EVFile = EVFile, filesetName = '038-120-200')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
