% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVCloseFile}
\alias{EVCloseFile}
\title{Close an open Echoview file (.EV)}
\usage{
EVCloseFile(EVFile)
}
\arguments{
\item{EVFile}{An Echoview file COM object}
}
\value{
a list object with two elements.  $chk: Boolean check indicating if the file was successfully closed; $msg: message for processing log.
}
\description{
This function closes an open Echoview file (.EV) via COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVCloseFile(EVFile)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
