% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EVCreateEditableLine}
\alias{EVCreateEditableLine}
\title{Create an EV Editable Line from an existing line
This function creates  an EV Line in an EV file object by name using COM scripting.}
\usage{
EVCreateEditableLine(EVFile, lineNameToCopy, editableLineName = NULL,
  Multiply = 1, Add = 0, SpanGaps = FALSE)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{lineNameToCopy}{a string containing the name of the line to copy}

\item{editableLineName}{=NULL desired editable line name.If null, EDIT is appended to lineNameToCopy and used as editable line name.}

\item{Multiply}{=1 multiply source line depth when creating new editable line}

\item{Add}{=0 add to source line depth when creating new editable line}

\item{SpanGaps}{=FALSE span gaps when creating new editable line}
}
\value{
an Echoview editable line object
}
\description{
Create an EV Editable Line from an existing line
This function creates  an EV Line in an EV file object by name using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj, '~~KAOS/KAOStemplate.EV')$EVFile

EVCreateEditableLine(EVFile = EVFile,
    lineNameToCopy='Offset seabed',
    editableLineName='Editable Seabed line')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVFindLineByName}} \code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}

