% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVCreateNew}
\alias{EVCreateNew}
\title{Create a new Echoview (.EV) file and adds raw data files to it}
\usage{
EVCreateNew(EVAppObj, templateFn = NULL, EVFileName, filesetName, dataFiles,
  CloseOnSave = TRUE)
}
\arguments{
\item{EVAppObj}{An EV application COM object arising from the call COMCreate('EchoviewCom.EvApplication')}

\item{templateFn}{= NULL Full path and filename of template file if used.}

\item{EVFileName}{Full path and filename of Echoview (.EV) file to be created.}

\item{filesetName}{Echoview fileset name}

\item{dataFiles}{vector of full path and name for each data file.}

\item{CloseOnSave}{= TRUE close the EV file in \code{EVFileName} once saved.}
}
\value{
a list object with two elements.  $EVFile: EVFile COM object for the newly created Echoview file, and $msg: message for processing log.
}
\description{
This function creates a new Echoview (.EV) file and adds raw data files to it via COM scripting.  Works well when populating an existing Echoview template file with raw data files.  The newly created Echoview file will remain open in Echoview and can be accessed via the $EVFile objected returned by a successful call of this function.
}
\details{
For the example code to run, the example data must be downloaded from.  NB the example code assumes the data and directory structure of the example data has been maintained.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
pathAndFn=list.files("~~/KAOS/raw/", full.names=TRUE)
#remove any evi type files
eviLoc=grep('.evi',pathAndFn)
if(length(eviLoc)>0) (pathAndFn=pathAndFn[-eviLoc])
EVCreateNew(EVAppObj=EVAppObj,
                 templateFn="~~/KAOS/KAOStemplate.EV",
                 EVFileName='~~/KAOS/kaos.ev',
                 filesetName="038-120-200",
                 dataFiles=pathAndFn, 
                 CloseOnSave = TRUE)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVNewFile}}  \code{\link{EVAddRawData}}  \code{\link{EVCloseFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
