% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVDeleteLine}
\alias{EVDeleteLine}
\title{Deletes an Echoview line object}
\usage{
EVDeleteLine(EVFile, evLine)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{evLine}{an Echoview line object}
}
\value{
a list object with one element- function message for log
}
\description{
This function deletes an Echoview line object using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

testline <- EVNewFixedDepthLine(EVFile = EVFile, depth = 50, lineName = "test_line")
EVDeleteLine(EVFile = EVFile, evLine = testline)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVNewFixedDepthLine}} \code{\link{EVFindLineByName}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
