% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVDeleteRegionClass}
\alias{EVDeleteRegionClass}
\title{Delete an Echoview region class}
\usage{
EVDeleteRegionClass(EVFile, regionClassCOMObj)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{regionClassCOMObj}{An Echoview region object}
}
\value{
a list object with two elements. $EVVar: An Echoview acoustic variable object, and $msg: message for processing log.
}
\description{
This function deletes a region class within an Echoview object using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
exampleClass <- EVRegionClassFinder(EVFile, "003")$regionClass
EVDeleteRegionClass(EVFile, exampleClass)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVRegionClassFinder}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
