% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVExportEchogramToImage.R
\name{EVExportEchogramToImage}
\alias{EVExportEchogramToImage}
\title{Export an Echoview acoustic variable to an image file}
\usage{
EVExportEchogramToImage(EVFile, acoVarName, filePath, heightInPixels,
  pingRange = c(-1, -1))
}
\arguments{
\item{EVFile}{Echoview file COM object}

\item{acoVarName}{The acoustic variable name to export}

\item{filePath}{Path and filename for the resulting image file.  The image type is controlled here by specifying the desired extension, e.g. *.jpg}

\item{heightInPixels}{Vertical resolution of exported image in pixels.  NB Echoview requires a 25 pixel minimum vertical resolution}

\item{pingRange}{=c(-1,1) Optional ping range over which to export the Echoview line object.  Default is to export the lien for all pings}
}
\value{
List of length two [1] Boolean flag indicating success or failure of export, and [1] message vector
}
\description{
This function exports an Echoview acoustic variable to an image file.
}
\details{
The following image types are available: *.bmp, *.jpg, *.jpeg, *.png, *.tif, *.tiff.
}
