% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVExportIntegrationAllRegions.R
\name{EVExportIntegrationAllRegions}
\alias{EVExportIntegrationAllRegions}
\title{Export echo  integration from Echoview for all regions in an Echoview acoustic variable}
\usage{
EVExportIntegrationAllRegions(EVFile, acoVarName, grid = FALSE, exportFn)
}
\arguments{
\item{EVFile}{("COMIDispatch) An Echoview file COM object}

\item{acoVarName}{(string) Echoview acoustic variable name}

\item{grid}{= FALSE (boolean) Integrate all data within a region to the region extents or when TRUE integrate on to a grid within the Echoview acoustic variable}

\item{exportFn}{CSV export file path and file name.}
}
\value{
list $msg messages resulting from function call
}
\description{
This function integrates all acoustic data within all regions for an Echoview acoustic variable.  The data can either be exported on a grid or as one integration per region.
}
\seealso{
\code{\link{EVChangeVariableGrid}}
}
