% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVExportIntegrationByCells}
\alias{EVExportIntegrationByCells}
\title{Export integration by cells for an acoustic variable}
\usage{
EVExportIntegrationByCells(EVFile, variableName, filePath)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{variableName}{a string containing the name of an EV acoustic variable}

\item{filePath}{a string containing the file path and name to save the exported data to}
}
\value{
a list object with 1 element: message for progessing log
}
\description{
This function exports the integration by cells for an acoustic variable using COM scripting. Note: This function will only work if the acoustic variable has a grid.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVExportIntegrationByCells(EVFile = EVFile, 
variableName = '38 seabed and surface excluded', 
filePath = '~~/KAOS/EVExportIntegrationByCells_example.csv')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
