% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVExportIntegrationByRegionByCells}
\alias{EVExportIntegrationByRegionByCells}
\title{Export integration by region name by cells for an acoustic variable}
\usage{
EVExportIntegrationByRegionByCells(EVFile, variableName, regionName, filePath)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{variableName}{a string containing the name of an EV acoustic variable}

\item{regionName}{a string containing the name of the region to export. Note: Input a single region's name, not region a class. To export by class, see EVIntegrationByRegionsByCellsExport.}

\item{filePath}{a string containing the file path and name to save the exported data to}
}
\description{
This function exports the integration by region by cells for an acoustic variable using COM scripting. Unlike EVIntegrationByRegionsByCellsExport, this function exports integration data for only for a single region name, not a region class. Note: This function will only work if the acoustic variable has a grid.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVExportIntegrationByRegionByCells(EVFile = EVFile, 
variableName = '38 seabed and surface excluded', 'Region1', 
filePath = '~~/KAOS/EVExportIntegrationByRegionByCells_example.csv')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
