% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVExportLine.R
\name{EVExportLineAsCSV}
\alias{EVExportLineAsCSV}
\title{Export an Echoview line as a CSV file}
\usage{
EVExportLineAsCSV(EVFile, acoVarName, lineName, filePath, pingRange = c(-1,
  -1))
}
\arguments{
\item{EVFile}{Echoview file COM object}

\item{acoVarName}{The acoustic variable name to which the line is associated}

\item{lineName}{Echoview line object name}

\item{filePath}{Path and filename for the resulting CSV file.}

\item{pingRange}{=c(-1,1) Optional ping range over which to export the Echoview line object.  Default is to export the lien for all pings}
}
\description{
This function exports an Echoview line object as a CSV file.
}
\details{
The Echoview line object should be associated with an Echoview acoustic variable object.  Optionally, the line can be exported for a subset of pings.
}
