% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVExportRegionDef}
\alias{EVExportRegionDef}
\title{Exports an Echoview region definition}
\usage{
EVExportRegionDef(EVFile, regionName, filePath)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{regionName}{a string containing the region name to export definitions for}

\item{filePath}{a string containing the name and file path of the file to export to}
}
\value{
a list object with one element- fucntion message for log
}
\description{
This function exports a single region's definition as a .csv file using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

EVExportRegionDef(EVFile, regionName = "Region1", filePath = "~~/KAOS/EVExportRegionDef.csv")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
