% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVExportRegionDefByClass}
\alias{EVExportRegionDefByClass}
\title{Exports definitions for all Echoview regions in a region class}
\usage{
EVExportRegionDefByClass(evRegionClass, filePath)
}
\arguments{
\item{evRegionClass}{an Echoview region class object}

\item{filePath}{a string containing the name and file path of the file to export to}
}
\value{
a list object with one element- fucntion message for log
}
\description{
This function exports definitions for all Echoview regions within a region class using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

regionClass <- EVRegionClassFinder(EVFile, "aggregations")$regionClass
EVExportRegionDefByClass(evRegionClass = regionClass, filePath = "~~/KAOS/EVExportRegionDefByClass_example.csv")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVRegionClassFinder}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
