% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVExportRegionSv}
\alias{EVExportRegionSv}
\title{Export Sv data for an Echoview acoustic variable by region}
\usage{
EVExportRegionSv(EVFile, variableName, regionName, filePath)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{variableName}{Echoview variable name for which to extract the data}

\item{regionName}{Echoview region name for which to extract the data}

\item{filePath}{File path and name (.csv) to save the data}
}
\value{
A list object with one element. $msg message for processing log
}
\description{
This function exports the Sv values as a .csv file for an acoustic variable by region using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVImportRegionDef(EVFile = EVFile, evrFile = '~~/KAOS/off transect regions/20030114_1200000000.evr', regionName = 'region_1')
EVExportRegionSv(EVFile = EVFile, variableName = '120 seabed and surface excluded', regionName = 'region_1', filePath = '~~/KAOS/EVExportRegionSv_example.csv')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVImportRegionDef}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
