% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EVExportUnderlying}
\alias{EVExportUnderlying}
\title{Export underlying data for an acoustic variable
This function exports underlying data for an Echoview acoustic variable using COM scripting}
\usage{
EVExportUnderlying(EVFile, variableName, pingRange = c(-1, -1), filePath)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{variableName}{a string containing the name of an EV acoustic variable}

\item{filePath}{a string containing the file path and name to save the exported data to}

\item{pingRange}{= c(-1,-1) ping range to export}
}
\value{
a list object with 1 element: message for progessing log
}
\description{
Export underlying data for an acoustic variable
This function exports underlying data for an Echoview acoustic variable using COM scripting
}
\details{
pingRange defaults are all pings i.e. start=-1 and stop=-1
}
\examples{
\dontrun{
EVAppObj = COMCreate('EchoviewCom.EvApplication')
EVFile = EVOpenFile(EVAppObj, '~~KAOS/KAOStemplate.EV')$EVFile
EVExportUnderlying(EVFile = EVFile, variableName = '38 seabed and surface excluded', pingRange = c(1, 100), filePath = '~~Desktop/test.csv')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}

