% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVExportWBST_FR.R
\name{EVExportWBST_FR}
\alias{EVExportWBST_FR}
\title{Export Wideband Single Targets Frequency Response}
\usage{
EVExportWBST_FR(
  EVFile = EVFile,
  STVar,
  outfn,
  AverageResults = 0,
  MinMax = 0,
  WindowSize = 1,
  WindowUnit = 2,
  EVVar
)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{STVar}{An Wideband Single Target Detection Echoview Variable, accepts inputs as Character, list or Variable object (COMIDispatch)}

\item{outfn}{FIlename for the TS export}

\item{MinMax}{Include Min and Max}

\item{WindowSize}{Set the WIndow Size}

\item{WindowUnit}{Unit of the window 0 = Meters, 1 = Pulse Lengths, 2 = Samples}

\item{EVVar}{Specifies a TS pulse compressed wideband variable for frequency response calculations.}
}
\description{
This function exports the broadband single targets frequency response
}
\examples{
#To be added - Needs Example data
#Starting Echoview
echoview = StartEchoview()
#Create a new EV File
EVFile <- EVCreateNew(EVAppObj=echoview, dataFiles = "rawfile.raw")$EVFile
#Set the First and second Operand
EVVar <- "Fileset 1: TS pulse compressed wideband pings T1"
Operand2 <- "Fileset 1: angular position pulse compressed wideband pings T1"
#Create a wideband single target detection varable
ST <- EVSTWideband(EVFile=EVFile,EVVar=EVVar,Operand2=Operand2)
# Change TS detection threshold
ST <- EVSTWideband(EVFile=EVFile,EVVar=EVVar,Operand2=Operand2,TsThreshold=-90)
#Export the single target frequency response
EVExportWBST_FR(ST,"WB_ST_FR.csv",0,0,1,2,EVVar)

}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
