% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVFilesInFileset}
\alias{EVFilesInFileset}
\title{Find names of all raw files in a fileset}
\usage{
EVFilesInFileset(EVFile, filesetName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{filesetName}{An Echoview fileset name}
}
\value{
A character vector containing the names of all .raw files in the fileset
}
\description{
This function returns the names of all .raw files in a fileset using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
file.names <- EVFilesInFileset(EVFile = EVFile, filesetName = '038-120-200')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
