% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVFindFilesetByName}
\alias{EVFindFilesetByName}
\title{Find an Echoview fileset in an Echoview file}
\usage{
EVFindFilesetByName(EVFile, filesetName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{filesetName}{Echoview fileset name to find}
}
\value{
a list object with two elements.  $fileset: found fileset COM object, and $msg: message for processing log.
}
\description{
This function finds an Echoview fileset in an Echoview file via COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVFileset <-EVFindFilesetByName(EVFile = EVFile, filesetName = '038-120-200')$filesetObj
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVNewFile}}  \code{\link{EVCreateNew}} \code{\link{EVCreateFileset}} \code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
