% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVFindFilesetTime}
\alias{EVFindFilesetTime}
\title{Find the time and date of the start and end of an Echoview fileset}
\usage{
EVFindFilesetTime(EVFile, filesetName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{filesetName}{An Echoview fileset name}
}
\value{
A list object with two elements $start.time: The date and time of the first measurement in the fileset, and $end.time: The date and time of the last measurement in the fileset
}
\description{
This function finds the date and time of the first and last measurement in a fileset using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
survey.time = EVFindFilesetTime(EVFile = EVFile, filesetName = '038-120-200')
start.time <- survey.time$start.time
end.time <- survey.time$end.time
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
