% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVFindLineByName}
\alias{EVFindLineByName}
\title{Find an EV Line object by name}
\usage{
EVFindLineByName(EVFile, lineName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{lineName}{a string containing the name of the line to find}
}
\value{
an Echoview line object
}
\description{
This function finds an EV Line in an EV file object by name using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVLine <- EVFindLineByName(EVFile = EVFile, lineName = "Fixed depth 250 m")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
