% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVFindRegionByName}
\alias{EVFindRegionByName}
\title{Finds an Echoview region by name}
\usage{
EVFindRegionByName(EVFile, regionName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{regionName}{a string containing the name of the region to find}
}
\value{
the Echoview region object
}
\description{
This function finds an Echoview region by name using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

testRegion <- EVFindRegionByName(EVFile, "Region1")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
