% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVFindRegionClass}
\alias{EVFindRegionClass}
\title{Find the class of an Echoview region object}
\usage{
EVFindRegionClass(evRegion)
}
\arguments{
\item{evRegion}{and Echoview Region object}
}
\value{
a string containing the class of the region
}
\description{
This function finds the class of an Echoview region object using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

ev.region <- EVFindRegionByName(EVFile, "Region1")
EVFindRegionClass(ev.region)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVFindRegionByName}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
