% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVFishTracking.R
\name{EVFishTracking}
\alias{EVFishTracking}
\title{Detect Fish Tracks}
\usage{
EVFishTracking(
  EVFile,
  EVVar,
  FishTrackRegionClass = "Unclassified regions",
  export = NULL,
  DataDimensions = NULL,
  Alpha.MajorAxis = NULL,
  Alpha.MinorAxis = NULL,
  Alpha.Range = NULL,
  Beta.MajorAxis = NULL,
  Beta.MinorAxis = NULL,
  Beta.Range = NULL,
  ExclusionDistance.MajorAxis = NULL,
  ExclusionDistance.MinorAxis = NULL,
  ExclusionDistance.Range = NULL,
  MissedPingExpansion.MajorAxis = NULL,
  MissedPingExpansion.MinorAxis = NULL,
  MissedPingExpansion.Range = NULL,
  Weights.MajorAxis = NULL,
  Weights.MinorAxis = NULL,
  Weights.Range = NULL,
  Weights.TS = NULL,
  Weights.PingGap = NULL,
  MinimumTargets = NULL,
  MinimumPings = NULL,
  MaximumGap = NULL
)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{EVVar}{An Echoview single target variable as a character string}

\item{FishTrackRegionClass}{string indicating the region class name for the fish tracks to be assigned}

\item{export}{full file path as a text string for an export file (i.e. "C:/FishTracking/Exports/FishTracks.csv"). NULL value skips export of fish tracks.}

\item{DataDimensions}{numeric Determine which algorithim is used for fish tracks. Accepted values are 2 (single or dual beam) or 4 (split beam)}

\item{Alpha.MajorAxis}{numeric Increases Alpha gain setting of Major Axis}

\item{Alpha.MinorAxis}{numeric Increases Alpha gain setting of Minor Axis}

\item{Alpha.Range}{numeric Increases Alpha gain setting of Range}

\item{Beta.MajorAxis}{numeric Increases Beta gain setting of Major Axis}

\item{Beta.MinorAxis}{numeric Increases Beta gain setting of Minor Axis}

\item{Beta.Range}{numeric Increases Beta gain setting of Range}

\item{ExclusionDistance.MajorAxis}{numeric Increases distance value of major axis in target gate parameters}

\item{ExclusionDistance.MinorAxis}{numeric Increases distance value of minor axis in target gate parameters}

\item{ExclusionDistance.Range}{numeric Increases distance value of range in target gate parameters}

\item{MissedPingExpansion.MajorAxis}{numeric Increases expansion value of major axis in target gate parameters}

\item{MissedPingExpansion.MinorAxis}{numeric Increases expansion value ofminor axis in target gate parameters}

\item{MissedPingExpansion.Range}{numeric Increases expansion value of missed pings range in target gate parameters. 0 disables missed ping expansion}

\item{Weights.MajorAxis}{numeric 1-100 Assigns the weighting of the major axis}

\item{Weights.MinorAxis}{numeric 1-100 Assigns the weighting of the minor axis}

\item{Weights.Range}{numeric 1-100 Assigns the weighting of the range}

\item{Weights.TS}{numeric 1-100 Assigns the weighting of the target strength}

\item{Weights.PingGap}{numeric 1-100}

\item{MinimumTargets}{numeric Minimum number of single targets to be considered a fish track}

\item{MinimumPings}{numeric Maximum number of pings in a track}

\item{MaximumGap}{numeric Maximum ping gap between tracks}
}
\description{
This function runs fish track detection on an single target echogram
Fish track algorithms and settings are explained in the [Echoview Help File]('https://support.echoview.com/WebHelp/Echoview.htm')
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVFishTracking(EVFile = EvFile,
EVVar = "Single target detection - wideband 1",
FishTrackRegionClass = "FishTracks",
MinimumTargets = 5,
export = "C:/FishTracking/Exports/FishTracks.csv")
}

}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
