% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVGetCalibrationFileName.R, R/EchoviewR.R
\name{EVGetCalibrationFileName}
\alias{EVGetCalibrationFileName}
\title{Get Name of ECS File for an EV File}
\usage{
EVGetCalibrationFileName(EVFile, filesetName)

EVGetCalibrationFileName(EVFile, filesetName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{filesetName}{a string containing the name of the Echoview fileset}
}
\value{
an object: returns the calibration file name
}
\description{
This reads the name and path of the ECS file for a given
EV file.

This function gets the calibration file name of a filesset using COM scripting
}
\details{
None to mention
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVGetCalibrationFileName(EVFile = EVFile, filesetName = "038-120-200")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
