% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVgetCalParValues.R
\name{EVgetCalParValues}
\alias{EVgetCalParValues}
\title{Get calibration settings pf an Echoview acoustic variable using COM}
\usage{
EVgetCalParValues(
  EVFile,
  acoVarName,
  pingNumber = 1,
  type = c("default", "set", "used"),
  verbose = FALSE
)
}
\arguments{
\item{acoVarName}{(character) Name of Echoview single target virtual variable}

\item{pingNumber}{= 0 (numeric) Ping number from which to extract calibration values.}

\item{type}{= =c('default','set','used') (character) Calibration value type.}

\item{verbose}{= FALSE (boolean) Should the calibration values be displayed in the console.}
}
\description{
This function gets the calibration values of an Echoview acousic variable using COM.
}
\details{
The function can read three types of calibration values "default", "set" and "used".
'@param EVFile ("COMIDispatch) An Echoview file COM object
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVAcoVarNameFinder}}
}
