% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVGetVariables.R
\name{EVGetVariables}
\alias{EVGetVariables}
\title{Get List of all variables}
\usage{
EVGetVariables(EVFile)
}
\arguments{
\item{EVFile}{An Echoview file COM object}
}
\description{
This function creates a list of all variables in the EV files
}
\examples{
#To be added - Needs Example data
#Starting Echoview
echoview = StartEchoview()
#Create a new EV File
EVFile <- EVCreateNew(EVAppObj=echoview, dataFiles = "rawfile.raw")$EVFile
Varname <- "Fileset 1: Sv pulse compressed wideband pings T1"
bottom <- EVBottomDetection(EVFile, EVVar=Varname, LineName="Bottom")
# Change the algorithm to Best bottom candidate
bottom <- EVBottomDetection(EVFile, EVVar=Varname, LineName="Bottom",algorithm=2)
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
