% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVImportLine}
\alias{EVImportLine}
\title{Imports an Echoview Line file}
\usage{
EVImportLine(EVFile, pathAndFn = NULL, lineName = NULL)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{pathAndFn}{string path and filename to .evl file.}

\item{lineName}{= NULL optional line name for imported line.}
}
\value{
a list object with two elements- [[1]] imported EV line COM object, and [[2]] function message for log
}
\description{
This function imports an Echoview line file (.evl) using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

#Example 1 - import a .evl line file
EVImportLine(EVFile, pathAndFn = '~~/KAOS/lineKAOS.evl')

#Example 2 - rename the imported line
EVImportLine(EVFile, pathAndFn='~~/KAOS/lineKAOS.evl', lineName='Line6' ) 
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}

