% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EVImportRegionDef}
\alias{EVImportRegionDef}
\title{Import an Echoview region definitions file (.evr)}
\usage{
EVImportRegionDef(EVFile, evrFile, regionName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{evrFile}{An Echoview region definitions file (.evr) path and name}

\item{regionName}{The name of the Echoview region}
}
\value{
A list object with one element. $msg message for processing log
}
\description{
This function imports a region definitions file (.evr) using COM scripting
}
\examples{
\dontrun{EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVImportRegionDef(EVFile = EVFile, evrFile = '~~/KAOS/off transect regions/20030114_1200000000.evr', regionName = 'region_1')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}

