% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/EVIntegrationByRegionsByCellsExport_Use_EV_thresholds.R
\name{EVIntegrationByRegionsByCellsExport_Use_EV_thresholds}
\alias{EVIntegrationByRegionsByCellsExport_Use_EV_thresholds}
\title{Export integration by regions by cells from an Echoview acoustic variable}
\usage{
EVIntegrationByRegionsByCellsExport_Use_EV_thresholds(EVFile, acoVarName,
  regionClassName, exportFn)
}
\arguments{
\item{EVFile}{An Echoview file object}

\item{acoVarName}{A string containing the name of an Echoview acoustic variable}

\item{regionClassName}{A string containing the name of an Echoview region class}

\item{exportFn}{export filename and path}
}
\value{
a list object with one element, $msg: message for processing log.
}
\description{
This function performs integration by regions by cells for a specified region class and exports the results using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVIntegrationByRegionsByCellsExport(EVFile, "120 aggregations", "aggregations", exportFn = "~~/KAOS/EVIntegrationByRegionsExport_example.csv")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
