% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVIntegrationByRegionsExport}
\alias{EVIntegrationByRegionsExport}
\title{Export integration by regions from an Echoview acoustic variable}
\usage{
EVIntegrationByRegionsExport(EVFile, acoVarName, regionClassName, exportFn,
  dataThreshold = NULL)
}
\arguments{
\item{EVFile}{An Echoview file object}

\item{acoVarName}{A string containing the name of an Echoview acoustic variable}

\item{regionClassName}{A string containing the name of an Echoview region class}

\item{exportFn}{export filename and path}

\item{dataThreshold}{An optional data threshold for export}
}
\value{
a list object with one element, $msg: message for processing log.
}
\description{
This function performs integration by regions and exports the results using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVIntegrationByRegionsExport(EVFile, "120 aggregations", "aggregations", exportFn = "~~/KAOS/EVIntegrationByRegionsExport_example.csv")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
