% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVmaxThresholdSet.R
\name{EVmaxThresholdSet}
\alias{EVmaxThresholdSet}
\title{Sets maximum data threshold for a variable object}
\usage{
EVmaxThresholdSet(EVFile, acoVarName, thres, colRng = 36)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{acoVarName}{(character) Echoview acoustic variable name}

\item{thres}{The new maximum threshold to be set (dB)}

\item{colRng}{maximum colour range possible for display settings.  Set to NA if you want any colour range.}
}
\value{
a list object with two elements. $thresholdSettings: The new threshold settings; $msg vector of messages
}
\description{
This function sets the maxmimum data threshold
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
varObj <- EVAcoVarNameFinder(EVFile, "38 seabed and surface excluded")$EVVar
EVmaxThresholdSet(varObj, -30)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVminThresholdSet}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
