% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVminThresholdSet}
\alias{EVminThresholdSet}
\title{Sets minimum data threshold for a variable object}
\usage{
EVminThresholdSet(varObj, thres)
}
\arguments{
\item{varObj}{An Echoview variable object}

\item{thres}{The new threshold to be set}
}
\value{
a list object with one element. $thresholdSettings: The new threshold settings
}
\description{
This function sets the minimum data threshold
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
varObj <- EVAcoVarNameFinder(EVFile, "38 seabed and surface excluded")$EVVar
EVminThresholdSet(varObj, -80)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVAcoVarNameFinder}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
