% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVNewAcousticVar}
\alias{EVNewAcousticVar}
\title{Add a new acoustic variable}
\usage{
EVNewAcousticVar(EVFile, oldVarName, enum)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{oldVarName}{a string containing the name of the acoustic variable to base the new variable on}

\item{enum}{Enum code for operator. See Echoview help file on EOperator for enum codes.}
}
\value{
an object: returns the new variable
}
\description{
This function adds a new acoustic variable using COM scripting
}
\examples{
\dontrun{
#create a 7x7 convolution of a variable
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVNewAcousticVar(EVFile = EVFile, oldVarName = "38 seabed and surface excluded", enum = 43)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
