% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVNewFile}
\alias{EVNewFile}
\title{Create a new Echoview file (.EV)}
\usage{
EVNewFile(EVAppObj, templateFn = NULL)
}
\arguments{
\item{EVAppObj}{An EV application COM object arising from the call COMCreate('EchoviewCom.EvApplication')}

\item{templateFn}{full path and filename for an Echoview template}
}
\value{
a list object with two elements.  $EVFile: EVFile COM object for the newly created Echovie file, and $msg: message for processing log.
}
\description{
This function creates a new Echoview file (.EV) via COM scripting which may be created from a template file if available
}
\examples{
\dontrun{
EVAppObj=COMCreate('EchoviewCom.EvApplication')
EVFile=EVNewFile(EVAppObj)$EVFile
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
