% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVNewFixedDepthLine}
\alias{EVNewFixedDepthLine}
\title{Creates a new fixed depth Echoview line}
\usage{
EVNewFixedDepthLine(EVFile, depth, lineName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{depth}{an integer specifying the fixed depth of the new line}

\item{lineName}{a string containing the name for the new line}
}
\value{
returns the EV Line object
}
\description{
This function creates a new Echoview line at a fixed depth using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

#create a new line at 50m depth named "testline"
newLine <- EVNewFixedDepthLine(EVFile = EVFile, depth = 50, lineName = "testline")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
