% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVNewLineRelativeRegion}
\alias{EVNewLineRelativeRegion}
\title{Creates a new line relative region in the current variable}
\usage{
EVNewLineRelativeRegion(EVFile, varName, regionName, line1, line2,
  firstPing = NA, lastPing = NA)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{varName}{a string containing the name of the acoustic variable to create the region in}

\item{regionName}{a string containing the name to assign to the new region}

\item{line1}{a string containing the name of the upper line limit}

\item{line2}{a string containing the name of the lower line limit}

\item{firstPing}{an optional integer for the ping to begin the region at}

\item{lastPing}{an optional integer for the ping to end the region at}
}
\value{
returns the EV Region object
}
\description{
This function creates a new line relative region in the current variable using COM scripting. The upper and lower depths are specified using Echoview line objects (these must already exist). Left and right bounds are optionally specified using ping number.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

#create a region between pings 1 - 100 and depths 20-250m
newRegion <- EVNewLineRelativeRegion(EVFile, 
"38 seabed and surface excluded", "test", 
"Fixed depth 6 m", "Fixed depth 250 m", 1, 100)

#create an unbounded region between depths 250-750m
newRegion <- EVNewLineRelativeRegion(EVFile, "38 seabed and surface excluded", "test", "Fixed depth 6 m", "Fixed depth 250 m")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
