% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVNewRegionClass}
\alias{EVNewRegionClass}
\title{Create a new Echoview region class}
\usage{
EVNewRegionClass(EVFile, className)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{className}{The name of the new Echoview region class}
}
\value{
A list object with one element. $msg message for processing log
}
\description{
This function creates a new region class using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVNewRegionClass(EVFile = EVFile, className = 'test_class')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
