% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVOpenFile}
\alias{EVOpenFile}
\title{Open an existing Echoview file (.EV)}
\usage{
EVOpenFile(EVAppObj, fileName)
}
\arguments{
\item{EVAppObj}{An EV application COM object arising from the call COMCreate('EchoviewCom.EvApplication')}

\item{fileName}{An Echoview file path and name.}
}
\value{
a list object with two elements.  $EVFile: EVFile COM object, and $msg: message for processing log.
}
\description{
This function opens an existing Echoview (.EV) file using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
