% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVPingSubset.R
\name{EVPingSubset}
\alias{EVPingSubset}
\title{Control an Echoview ping subset virtual variable using COM scripting.}
\usage{
EVPingSubset(EVFile, acoVarName, pingsubsetString, includePingRanges = TRUE)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{acoVarName}{Acoustic variable name}

\item{pingsubsetString}{A ping subset string (see examples).}

\item{includePingRanges}{=TRUE Boolean.  Default is for data within the ping range to be included}
}
\value{
a list object with two elements.  $success: ping subset successfully changed, and $msg: message for processing log.
}
\description{
This function opens an existing Echoview (.EV) file using COM scripting.
}
\examples{
\dontrun{
EVAppObj=COMCreate('EchoviewCom.EvApplication')
EVOpenFile(EVAppObj,'~\\\\example1.EV')
EVpingSubset(EVFile=EVFile,acoVarName='Ping subset 38',pingsubsetString='2-7')

}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
