% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVRegionClassFinder}
\alias{EVRegionClassFinder}
\title{Find an Echoview region class by name}
\usage{
EVRegionClassFinder(EVFile, regionClassName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{regionClassName}{A string containing the name of an Echoview region}
}
\value{
a list object with two elements. $regionClass: The class of the region, and $msg: message for processing log.
}
\description{
This function finds an echoview region class by name.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
aggregationsClass <- EVRegionClassFinder(EVFile, "aggregations")$regionClass
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
