% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVRenameAcousticVar.R
\name{EVRenameAcousticVar}
\alias{EVRenameAcousticVar}
\title{Rename an Echoview acoustic variable}
\usage{
EVRenameAcousticVar(EVFile, acoVarName, newName)
}
\arguments{
\item{EVFile}{Echoview file COM object}

\item{acoVarName}{The acoustic variable name to which a new variable will be added.}

\item{newName}{New name of the acoustic variable}
}
\value{
list with three elements: $varObj acoustic variable COM object; $objName new name of the variable object;$msg message vector
}
\description{
This function renames an existing acoustic variable
}
\note{
The function will check if the proposed new variable new already exists.  If the porposed new name exists, then the new name will be chnaged to 'newName duplicate name'.
}
