% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVRenameLine}
\alias{EVRenameLine}
\title{Renames an Echoview Line object}
\usage{
EVRenameLine(EVFile, evLine, newName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{evLine}{an Echoview line object}

\item{newName}{a string containing the new name for the line}
}
\value{
a list object with one element- fucntion message for log
}
\description{
This function renames an Echoview line object using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

testline <- EVNewFixedDepthLine(EVFile = EVFile, depth = 50, lineName = "testline")
EVRenameLine(EVFile = EVFile, evLine = testline, newName = "line40")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVNewFixedDepthLine}} \code{\link{EVFindLineByName}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
