% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVResampleByPingControl.R
\name{EVResampleByPingControl}
\alias{EVResampleByPingControl}
\title{Control the 'Resample by ping' Echoview virtual variable settings using COM}
\usage{
EVResampleByPingControl(
  EVFile,
  acoVarName,
  NumberOfSamples = NULL,
  NumberOfPings = NULL,
  StartRange = NULL,
  StopRange = NULL,
  verbose = FALSE
)
}
\arguments{
\item{EVFile}{Echoview file COM object}

\item{acoVarName}{name of the 'Resample by pings' acoustic variable}

\item{NumberOfSamples}{number of samples to export in the depth dimension.}

\item{NumberOfPings}{number of pings to resample}

\item{StartRange}{start range, m}

\item{StopRange}{maximum range, m}

\item{verbose}{= FALSE}
}
\value{
List of length three [1] $previousSettings [2] $currentSettings  [3] message vector
}
\description{
This function controls the 'Resample by ping' Echoview virtual variable settings using COM
}
