% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVSaveAsFile}
\alias{EVSaveAsFile}
\title{Perform save as operation on an open Echoview file (.EV)}
\usage{
EVSaveAsFile(EVFile, fileName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{fileName}{An Echoview file path and name.}
}
\value{
a list object with two elements.  $chk: Boolean check indicating if the file was successfully saved; $msg: message for processing log.
}
\description{
This function performs a save as operation on an existing Echoview (.EV) file using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVSaveAsFile(EVFile = EVFile, fileName = '~~/KAOS/KAOStemplate_test.EV')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVSaveFile}} \code{\link{EVCloseFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
