% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVSchoolsDetect}
\alias{EVSchoolsDetect}
\title{Schools Detection in Echoview}
\usage{
EVSchoolsDetect(EVFile, acoVarName, outputRegionClassName,
  deleteExistingRegions, distanceMode, maximumHorizontalLink,
  maximumVerticalLink, minimumCandidateHeight, minimumCandidateLength,
  minimumSchoolHeight, minimumSchoolLength, dataThreshold)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{acoVarName}{A string containing the name of the acoustic variable to perform the analysis on}

\item{outputRegionClassName}{A string containing the name of the output region}

\item{deleteExistingRegions}{Logical TRUE or FALSE}

\item{distanceMode}{for schools detection (see Echoview help).}

\item{maximumHorizontalLink}{The maximum horizontal link in meters}

\item{maximumVerticalLink}{The maximum vertical link in meters}

\item{minimumCandidateHeight}{The minimum candidate height in meters}

\item{minimumCandidateLength}{the minimum candidate length in meters}

\item{minimumSchoolHeight}{The minimum school height in meters}

\item{minimumSchoolLength}{The minimum school length in meters}

\item{dataThreshold}{minimum integration threshold (units: dB re 1m^-1)}
}
\value{
a list object with four elements. $nbrOfDetectedschools, $thresholdData, $schoolsSettingsData, and $msg: message for processing log
}
\description{
This function performs schools detection in Echoview using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVSchoolsDetect(EVFile = EVFile,
               acoVarName='120 7x7 convolution',
               outputRegionClassName = 'aggregations',
               deleteExistingRegions = TRUE,
               distanceMode = "GPS distance",
               maximumHorizontalLink = 15, #m
               maximumVerticalLink = 5,#m
               minimumCandidateHeight = 1, #m
               minimumCandidateLength = 10, #m
               minimumSchoolHeight = 2, #m
               minimumSchoolLength = 15, #m
               dataThreshold = -80)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
