% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVSchoolsDetSet}
\alias{EVSchoolsDetSet}
\title{Change schools detection settings}
\usage{
EVSchoolsDetSet(EVFile, varObj, distanceMode, maximumHorizontalLink,
  maximumVerticalLink, minimumCandidateHeight, minimumCandidateLength,
  minimumSchoolHeight, minimumSchoolLength)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{varObj}{the EV acoustic object to change schools detection parameters for}

\item{distanceMode}{which distance mode to use}

\item{maximumHorizontalLink}{maximum linking distance for a swarm}

\item{maximumVerticalLink}{maximum vertical linking distance for a school}

\item{minimumCandidateHeight}{minimum candidate height}

\item{minimumCandidateLength}{minimum candidate length}

\item{minimumSchoolHeight}{minimum school height}

\item{minimumSchoolLength}{minimum school length}
}
\value{
a list object with two elements.
}
\description{
This function changes schools detection settings for an acoustic variable using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
varObj <- EVAcoVarNameFinder(EVFile, "120 7x7 convolution")$EVVar
changeSettings <- EVSchoolsDetSet(EVFile, varObj, distanceMode = "GPS distance",
                                maximumHorizontalLink = 10,
                                maximumVerticalLink = 5,
                                minimumCandidateHeight = 2,
                                minimumCandidateLength = 3,
                                minimumSchoolHeight = 4,
                                minimumSchoolLength = 2)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVAcoVarNameFinder}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
