% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schDetectV12.R
\name{EVSchoolsDetSetV12}
\alias{EVSchoolsDetSetV12}
\title{Change schools detection settings for Echvoiew >=V12}
\usage{
EVSchoolsDetSetV12(
  EVApp,
  EVFile = NULL,
  varObj = NULL,
  distanceMode,
  maximumHorizontalLink,
  maximumVerticalLink,
  minimumCandidateHeight,
  minimumCandidateLength,
  minimumSchoolHeight,
  minimumSchoolLength
)
}
\arguments{
\item{EVApp}{= EV application COM object}

\item{EVFile}{=NULL not used An Echoview file COM object}

\item{varObj}{=NULL not used the EV acoustic object to change schools detection parameters for}

\item{distanceMode}{which distance mode to use}

\item{maximumHorizontalLink}{maximum linking distance for a swarm}

\item{maximumVerticalLink}{maximum vertical linking distance for a school}

\item{minimumCandidateHeight}{minimum candidate height}

\item{minimumCandidateLength}{minimum candidate length}

\item{minimumSchoolHeight}{minimum school height}

\item{minimumSchoolLength}{minimum school length}
}
\value{
a list object with two elements.
}
\description{
This function changes schools detection settings in Echoview version >=12
}
\examples{
\dontrun{
library(EchoviewR)
EVApp=NULL
EVApp=EchoviewR::StartEchoview()
EVFile=EchoviewR::EVOpenFile(EVAppObj = EVApp,
                            fileName='x:\\\\KAOS\\\\20120519_KAOS_all_2.EV')$EVFile

EVSchoolsDetSetV12(EVApp=EVApp,
distanceMode="GPSDistance",
maximumHorizontalLink=20,
maximumVerticalLink=10,
minimumCandidateHeight=10,
minimumCandidateLength=40,
minimumSchoolHeight=10,
minimumSchoolLength=50)

}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVSchoolsDetSet}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
