% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVsetExclusionLines.R
\name{EVsetExclusionLines}
\alias{EVsetExclusionLines}
\title{Change the top and bottom exclusion lines in an Echoview acoustic variable}
\usage{
EVsetExclusionLines(EVFile, acoVarName, newAboveExclusionLine = NULL,
  newBelowExclusionLine = NULL)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{acoVarName}{Acoustic variable name.}

\item{newAboveExclusionLine}{Name of the top exclusion line to be used. Default is NULL in which case the top exclusion line is unchanged.}

\item{newBelowExclusionLine}{Name of the bottom exclusion line to be used. Default is NULL in which case the bottom exclusion line is unchanged.}
}
\value{
a list object with two elements.  $analysisProperties: The analysis properties COM object specified in \code{acoVarName}, and $msg: message vector for the processing log.
}
\description{
This function changes the top and bottom exclusion lines in an Echoview acoustic variable using COM scripting.
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\link{EVAcoVarNameFinder}
}
\author{
Martin Cox
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
