% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVShiftRegionDepth}
\alias{EVShiftRegionDepth}
\title{Change the depth of an Echoview Region}
\usage{
EVShiftRegionDepth(EVFile, regionName, depthMultiply, depthAdd)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{regionName}{a string containing the name of the Echoview region}

\item{depthMultiply}{a numeric value to multiply the vertical size of the region by}

\item{depthAdd}{a numeric value to offset the region depth by. Positive = decrease depth; Negative = increase depth}
}
\value{
a list object with one element, $msg: message for processing log.
}
\description{
This function shifts the depth of an echoview region using COM scripting. Vertical size of the region and depth offset can be changed.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

#Double region vertical size and decrease depth by 100m
EVShiftRegionDepth(EVFile, "testregion", 2, 100)

#Triple region vertical size without changing depth offset
EVShiftRegionDepth(EVFile, "testregion", 3, 0)

#Change region depth offset by -50m without changing vertical size
EVShiftRegionDepth(EVFile, "testregion", 1, -50)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
