% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVShiftRegionTime}
\alias{EVShiftRegionTime}
\title{Change the time of an Echoview Region}
\usage{
EVShiftRegionTime(EVFile, regionName, days = 0, hours = 0, minutes = 0,
  seconds = 0, milliseconds = 0)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{regionName}{a string containing the name of the Echoview region}

\item{days}{an integer value specifying days to add (positive) or subtract (negative). Default = 0}

\item{hours}{an integer value specifying hours to add (positive) or subtract (negative). Default = 0}

\item{minutes}{an integer value specifying minutes to add (positive) or subtract (negative). Default = 0}

\item{seconds}{an integer value specifying seconds to add (positive) or subtract (negative). Default = 0}

\item{milliseconds}{an integer value specifying milliseconds to add (positive) or subtract (negative). Default = 0}
}
\value{
a list object with one element, $msg: message for processing log.
}
\description{
This function shifts the time of an echoview region using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile

#Shift the region time by 10 seconds
EVShiftRegionTime(EVFile, "testregion", seconds = 10)

#Subtract 1 hour from the region time
EVShiftRegionTime(EVFile, "testregion", hours = -1)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
