% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVsingleTargetPars.R
\name{EVsingleTargetPars}
\alias{EVsingleTargetPars}
\title{Change the single target settings for Echoview single target virtual variable}
\usage{
EVsingleTargetPars(EVFile, acoVarName, ExcludeAboveLine = NULL,
  ExcludeBelowLine = NULL, MinimumPulseLength = NULL,
  MaximumPulseLength = NULL, PLDL = NULL, TsThreshold = NULL,
  verbose = FALSE)
}
\arguments{
\item{EVFile}{("COMIDispatch) An Echoview file COM object}

\item{acoVarName}{(character) Name of Echoview single target virtual variable}

\item{ExcludeAboveLine}{(character) Shallower exclusion line}

\item{ExcludeBelowLine}{(character) Deeper exclusion line}

\item{MinimumPulseLength}{(numeric) Minimum pulse length}

\item{MaximumPulseLength}{(numeric) Maximum pulse length}

\item{PLDL}{(numeric) Pulse length determination level}

\item{TsThreshold}{(numeric) minimum target strength threshold (dB)}

\item{verbose}{= FALSE (boolean)}
}
\value{
list $settings 2 column data frame of pre-change post-change parameter values, $msg= vector of messages
}
\description{
Change the single target settings for Echoview virtual variable
}
\details{
This function will work with single target detection viritual variables
}
\note{
THis function has only been tested using Simrad EK60 split-beam data
}
