% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVSingleTargetsRegionsByCellsExport.R
\name{EVSingleTargetsRegionsByCells}
\alias{EVSingleTargetsRegionsByCells}
\title{Function to export TS frequency distribution.}
\usage{
EVSingleTargetsRegionsByCells(EVFile, TSacoVarName, regionClassName, exportFn)
}
\arguments{
\item{EVFile}{}

\item{TSacoVarName}{}

\item{regionClassName}{}

\item{exportFn}{}
}
\value{
This function exports single target by region-cells. The result is
a file with a row per interval-layer with targets meeting defined 
parameters, with the mean, minimum, and maximum TS for that cell.
}
\description{
This function runs region-by cell exports of TS frequency distribution 
for EV files identified in the data directory selected and exports them to 
the output directory selected. Can be called on its own but is also used in 
the function ExportSvSingleTargetsRegionbyCell.
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
