% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVsplitBeamSingleTargetPars.R
\name{EVsplitBeamSingleTargetPars}
\alias{EVsplitBeamSingleTargetPars}
\title{Change the SPLIT-BEAM single target settings for an Echoview split-beam single target virtual variable}
\usage{
EVsplitBeamSingleTargetPars(EVFile, acoVarName,
  BeamCompensationModel = "SimradLobe", MaximumBeamCompensation = NULL,
  MaximumStdDevOfMajorAxisAngles = NULL,
  MaximumStdDevOfMinorAxisAngles = NULL, verbose = FALSE)
}
\arguments{
\item{EVFile}{("COMIDispatch) An Echoview file COM object}

\item{acoVarName}{(character) Name of Echoview single target virtual variable}

\item{BeamCompensationModel}{='SimradLOBE' c('Simrad LOBE','BioSonics','HTI','Furuno','Sonic')Beam compensation model}

\item{MaximumBeamCompensation}{=NULL (numeric)}

\item{MaximumStdDevOfMajorAxisAngles}{=NULL (numeric) maximum standard deviation of Major axis angles (degrees)}

\item{MaximumStdDevOfMinorAxisAngles}{=NULL (numeric) maximum standard deviation of Minor axis angles (degrees)}

\item{verbose}{= FALSE (boolean)}
}
\value{
list $settings 2 column data frame of pre-change and post-change parameter values , $msg= vector of messages
}
\description{
Change the split-beam single target settings for Echoview virtual variable of type split beam (method 1) and split-eabm (method 2)/
}
\details{
This function will work with split-beam single target detection viritual variables (methods 1 and 2).
}
\note{
THis function has only been tested using Simrad EK60 split-beam data
}
