% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVSTWideband.R
\name{EVSTWideband}
\alias{EVSTWideband}
\title{Broadband Single target detection}
\usage{
EVSTWideband(
  EVFile,
  EVVar,
  Operand2,
  TsThreshold = NULL,
  MinimumPulseLength = NULL,
  MaximumPulseLength = NULL,
  BeamCompensationModel = NULL,
  MaximumBeamCompensation = NULL,
  MaximumStdDevOfMajorAxisAngles = NULL,
  MaximumStdDevOfMinorAxisAngles = NULL
)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{EVVar}{An Echoview Variable, accepts inputs as Character, list or Variable object (COMIDispatch)}

\item{Operand2}{An Echoview Variable, accepts inputs as Character, list or Variable object (COMIDispatch)}

\item{TsThreshold}{numeric [dB] Single target detection threshold}

\item{MinimumPulseLength}{numeric Minimum normalized pulse length}

\item{MaximumPulseLength}{numeric Maximum normalized pulse length ,}

\item{BeamCompensationModel}{numeric [1,4] 1 = Simrad Lobe, 4 = None,}

\item{MaximumBeamCompensation}{NULL=,numeric Maximum beam compensationngles = NULL,}

\item{MaximumStdDevOfMajorAxisAngles}{Maximum standard deviation of major axis angles}

\item{MaximumStdDevOfMinorAxisAngles}{Maximum standard deviation of minor axis angles}
}
\description{
This function creates a Wideband Single Target Detection Variable
}
\examples{
#To be added - Needs Example data
#Starting Echoview
echoview = StartEchoview()
#Create a new EV File
EVFile <- EVCreateNew(EVAppObj=echoview, dataFiles = "rawfile.raw")$EVFile
#Set the First and second Operand
EVVar <- "Fileset 1: TS pulse compressed wideband pings T1"
Operand2 <- "Fileset 1: angular position pulse compressed wideband pings T1"
#Create a wideband single target detection varable
ST <- EVSTWideband(EVFile=EVFile,EVVar=EVVar,Operand2=Operand2)
# Change TS detection threshold
ST <- EVSTWideband(EVFile=EVFile,EVVar=EVVar,Operand2=Operand2,TsThreshold=-90)

}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
