% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVVersion.R
\name{EVVersion}
\alias{EVVersion}
\title{Get the current version of Echoview}
\usage{
EVVersion(EVApp = NULL)
}
\arguments{
\item{EVApp}{An Echoview Application object}
}
\description{
This function returns the version of Echoview that was started through EchoviewR
}
\examples{
\dontrun{
#Create an Echoview application object and start Echoview
EVApp <- StartEchoview()
#get the current version of the Echoview App started through EchoviewR
EVVersion(EVApp)
#Quit the application again
QuitEchoview(EVApp)
}

}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
