% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{exportMIF}
\alias{exportMIF}
\title{Write a map info file for import into Echoview.}
\usage{
exportMIF(coords, pathAndFileName, pointNameExport = FALSE,
  pointNameScaleFactor = c(0.05, 0.03))
}
\arguments{
\item{coords}{a set of coordinates as a two column (longitude and latitude) matrix.}

\item{pathAndFileName}{character string MIF file export path and filename}

\item{pointNameExport}{Boolean (default=FALSE) export the name of each point.}

\item{pointNameScaleFactor}{=c(0.005,0.03) see details}
}
\value{
Nothing
}
\description{
This function writes a polygon MIF file for import into Echoview and will
typically be used to export survey line transects from, for example, 
\link{centreLawnOnPosition}.
}
\details{
Exporting the name of each point uses the row name in the \code{coords} argument.  

The \code{pointNameScaleFactor} argument is a two element numeric vector specifying the bounding 
box for each point name.  The bounding box is calculated for point name i by specifying the lower left hand corner 
of the point name position, with the upper right hand corner specified as 
c(lon_i,lat_i)+diff(range(lon_1..n,lat_1..n))*pointNameScaleFactor where n is the total number of points i.e nrow(coords)
}
\examples{
\dontrun{
coords=centreLawnOnPosition(centreLon=-33,centreLat=-55.5,lineLengthkm=100,lineSpacingkm=10,
startBearingdeg=30,numOfLines=50)
exportMIF(coords=coords,pathAndFileName='c:\\\\Users\\\\martin_cox\\\\Documents\\\\test4.mif')
}
}
