% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportSvTSdistRegionbyCell.R
\name{ExportSvTSdistRegionbyCell}
\alias{ExportSvTSdistRegionbyCell}
\title{Export Sv and TS distribution by region-cell from a folder of EV files.}
\usage{
ExportSvTSdistRegionbyCell(dat.dir, exp.dir, SvacoVarName, TSacoVarName,
  regionClassName)
}
\arguments{
\item{dat.dir}{}

\item{exp.dir}{}

\item{SvacoVarName}{}

\item{TSacoVarName}{}

\item{regionClassName}{}
}
\value{
This function generates csv files of Sv and TS frequency distribution data with the result 
being one of these files for each EV file processed.
}
\description{
This function runs region-by cell exports of Sv and TS frequency distribution 
for EV files identified in a single folder selected by the user and exports them to 
the output directory selected by the user. Can of course be used to process
multiple folders of data with additional code.
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
