% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varnames_in_EVfiles.R
\name{get_acoVarNames_in_EVfiles}
\alias{get_acoVarNames_in_EVfiles}
\title{Generate a list of acoustic variables in list of EV files and
save that list in a csv file in the folder containing the 
EV files.}
\usage{
get_acoVarNames_in_EVfiles(ev.dir, EVFile)
}
\arguments{
\item{ev.dir}{Character, the directory containing the EV files}

\item{EVFile}{Character, the name of the EV file being processed. Determine by
funciton.}
}
\value{
A csv file with variables "var", the name of each variable, "EV.name", the
full path/name for each EV file, "Nvar", the number of variables in the EV file,
and "EVfolder.name", the name of the subfolder containing the EV file.
}
\description{
This function generates a list of all variables present in each
EV file in the list of EV files generated by the directory choice.
}
\author{
Dave Warner
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
