% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRaws.R
\name{getRaws}
\alias{getRaws}
\title{Find Raw files based on Time and Date, providing folder containing all Raw files}
\usage{
getRaws(dir, StartDate, StartTime, EndDate, EndTime, tzraw = "UTC",
  tzinp = "UTC", timebuffer = 0, fileSlope = FALSE)
}
\arguments{
\item{dir}{= path [character]}

\item{StartDate}{[character]  Date under the shape of YYYYMMDD (example: 20171127)}

\item{StartTime}{[character]  Time under the shape of HHMMSS (example: 085617)}

\item{EndDate}{[character]  Date under the shape of YYYYMMDD (example: 20171127)}

\item{EndTime}{[character]  Time under the shape of HHMMSS (example: 085617)}

\item{tzraw}{="UTC"[character]  timezone under which the raw files were recorded e.g.UTC or Australia/Brisbane or AEST}

\item{tzinp}{="UTC"[character]  timezone of the entered start and end points e.g.UTC or Australia/Brisbane or AEST}

\item{timebuffer}{= 0 [numeric in seconds]  add a buffer around the times for which files should be selected}

\item{fileSlope}{= FALSE [boolean] If TRUE the raw file that partly falls within the start of the time interval.}
}
\value{
charater array with raw data filenames
}
\description{
This functions generates datetime object based on the raw data filenames and selects the required raw files according to 
a user defined start and end time
}
\note{
Dependencies: lubridate - Requires lubridate package for easy computation of time intervals
}
\examples{
\dontrun{
StartDate = 20170901
StartTime = "080917"
EndDate = 20170901
EndTime = 172356
rawdir = "F:\\\\RawFolder\\\\"
getRaws(dir,StartDate, StartTime,EndDate, EndTime, tzraw="UTC", tzinp="UTC")
 }
 
 
}
\keyword{Echoview}
\keyword{com}
\keyword{scripting}
