% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JollyAndHampton1990.R
\name{jhMultipleStrataF}
\alias{jhMultipleStrataF}
\title{Estimate mean and variance across multiple strata using the Jolly and Hampton (1990) method}
\usage{
jhMultipleStrataF(strata)
}
\arguments{
\item{strata}{list of multiple strata, each element must be the result of a call of \link{jhF}.  See details.}
}
\description{
Estimate mean and variance across multiple strata using the Jolly and Hampton (1990) method
}
\details{
The call of \link{jhF} used to create an element in the  \code{strata} argument must have the name of the stratum and the stratum area.
#'@references Jolly, G. M., and Hampton, I. 1990. A stratified random transect design for acoustic surveys of fish stocks. Canadian Journal of Fisheries and Aquatic Sciences, 47: 1282-1291.
}
\examples{
\dontrun{
svy1=jhF(transectLength=tLengths,
transectMeanDen=denV,
transectName=NULL,svyName='test',area=svyA)
svy2=jhF(transectLength=tLengths,
transectMeanDen=denV,
transectName=NULL,svyName='test',area=svyA)

jhMultipleStrataF(strata=list(svy1,svy2))

svy3=jhF(transectLength=tLengths,
transectMeanDen=denV,
transectName=NULL,svyName='test',area=svyA*2)

jhMultipleStrataF(strata=list(svy1,svy3))

svy4=jhF(transectLength=tLengths,
transectMeanDen=denV*2,
transectName=NULL,svyName='test',area=svyA)

jhMultipleStrataF(strata=list(svy1,svy4))

}
}
\seealso{
\link{jhF}
}
