% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEVRfromDT.R
\name{makeEVRfromDT}
\alias{makeEVRfromDT}
\title{Create an Echoview EVR (Region definition) file from depth and time intervals}
\usage{
makeEVRfromDT(depths, times, dates, rName = "Region", rClass = "Selection",
  rType = 1, dir = NULL, fn = NULL, rNotes = list(""))
}
\arguments{
\item{depths}{list of depths (can be multiple regions, where each is a list element)}

\item{times}{(character) list of times in HHMMSS}

\item{dates}{list of dates as character \%Y\%m\%d (can be multiple regions, where each is a list element)}

\item{rName}{= "Region" (character) Name of the region}

\item{rClass}{= "Selection".  character Define class, otherwise it will be Unclassified}

\item{rType}{= (integer; 0 to 4) see details}

\item{dir}{= NULL character Path to the output folder}

\item{fn}{= NULL character Name of the output filename}

\item{rNotes}{= list("") character Region Notes}
}
\value{
saves an Echoview EVR (Region definition) files
}
\description{
This function generates Echoview Region definitions according to predefined starting and end depts at given times.
The results are save as .EVR file for import into Echoview
}
\details{
The following region types (integer 0 to 4) are available in Echoview 0 - Bad Data (no data); 1 - Analysis; 2 - Marker; 3 - fishtracks; 4 - Bad Data (empty water)
}
\note{
This function is experimental
}
\examples{
\dontrun{
# Minimal example generating 2 regions
depths <- list(c(10,35,50,74,82), c(15,250,300,350,60))
times <-  list(c("090000","091000","091500","091000","090300"),
               c("093000","094000","094500","094000","093300"))
dates <- list(c(rep("20170901",5)),c(rep("20170901",5)))
dir <- 'F:\\\\'
fn <- 'imaginaryRegions'
makeEVRfromDT(depths,times,dates,fn=fn,dir=dir)
 }
}
\seealso{
\code{\link{makeRectRegion}}
}
\author{
Sven Gastauer
}
\keyword{Echoview}
\keyword{com}
\keyword{scripting}
