% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRectRegion.R
\name{makeRectRegion}
\alias{makeRectRegion}
\title{Make a rectangular Echoview region (EVR) file.}
\usage{
makeRectRegion(dateStart, timeStart, dateStop, timeStop, minDepth = 0,
  maxDepth, outPathFn, regionClassName, regionType, append = FALSE)
}
\arguments{
\item{dateStart}{date  stamp for the start of the region (yyymmdd)}

\item{timeStart}{time stamp for the start of the region (hhmmss)}

\item{dateStop}{date stamp for the end of the region (yyymmdd)}

\item{timeStop}{time stamp for the end of the region (hhmmss)}

\item{minDepth}{=0 Minimum depth for the region}

\item{maxDepth}{maximum depth for the region}

\item{outPathFn}{output parth and filename.}

\item{regionClassName}{region class name}

\item{regionType}{region Type (integer)}

\item{append}{= FALSE should be region be added to the EVR file given in outPathFn}
}
\value{
list: $msg messages from the function call $exportFLAG Boolean TRUE EVR file created or added.  FALSE otherwise
}
\description{
This function creates a rectangular (2D) Echoview region file.  The function can add regions to an existing EVR file.
}
\details{
Region types supported by Echoview: 0 = bad (no data); 1 = analysis; "2" = marker; 3 = fishtracks, and 4 = bad (empty water)
}
\examples{
\dontrun{makeRectRegion(dateStart='20170815',timeStart='001525',
dateStop='20170815',timeStop='001738', 
minDepth=20,maxDepth=100,
outPathFn='~/test2.evr',
regionClassName='ctd',append=FALSE)}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
