% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msDATEConversion.R
\name{msDATEConversion}
\alias{msDATEConversion}
\title{Convert a Microsoft DATE object to a human readable date and time}
\usage{
msDATEConversion(dateObj)
}
\arguments{
\item{dateObj}{a Microsoft date object}
}
\value{
time stamp in the format yyyy-mm-dd hh:mm:ss
}
\description{
Time stamps in Echoview, such as start and end times of, for example, invididual regions, use the Microsoft DATE format.  This function converts the Microsoft DATE object to a human readable date and time.  NB no time zone is returned.
}
\seealso{
www.echoview.com
}
