% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVQuitEchoview.R
\name{QuitEchoview}
\alias{QuitEchoview}
\title{Close an Echoview instance through COM scripting}
\usage{
QuitEchoview(EVAppObj)
}
\arguments{
\item{EVAppObj}{An Echoview Application COM object}
}
\value{
Termination of the application and removal of the Echoview instance object
}
\description{
This function opens an Echoview instance
}
\examples{
\dontrun{
EVAppObj=StartEchoview()
QuitEchoview(EVAppObj)

}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
