% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readevl2R.R
\name{readevl2R}
\alias{readevl2R}
\title{Read and Echoview line (.evl) file into R}
\usage{
readevl2R(pathFn, tz = "GMT")
}
\arguments{
\item{pathFn}{path and filename of evl file}

\item{tz}{='GMT' time zone see as.POSIXlt}
}
\value{
list object of class evl: [1]$timing ($start, $end start and end time stamps); [2] $line1 line 1 - evl file first line; 
[3] $n number of points in evl file; [4] $bodyEVL body of evl file; [5] $timeStamp time stamp for; [6] msg message log;;
[7] $pathFn path and filename of Echoview line file; [8] tz time zone
}
\description{
Read and Echoview line (.evl) file into R
}
