% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUnderlyingDataCSV.R
\name{readUnderlyingDataCSV}
\alias{readUnderlyingDataCSV}
\title{Read CSV resulting from a call of EVExportUnderlying()}
\usage{
readUnderlyingDataCSV(pathAndFn, samplePreFix = "sample", depthPreFix = "z",
  asList = FALSE, onlyHeader = FALSE, lastHeaderCol = 13)
}
\arguments{
\item{pathAndFn}{The path and filename of the CSV fileonlyHeader}

\item{samplePreFix}{='sample' (character)text to add to the sample column names}

\item{depthPreFix}{= 'z' (character) text to add to the sample row names (depth)}

\item{asList}{= FALSE (boolean) return the CSV as a list. See details.}

\item{onlyHeader}{= FALSE (boolean) return only the header of the CSV, ignored if asList = TRUE. See details.}

\item{lastHeaderCol}{= 13 The last column of the header information.}
}
\description{
This function reads a CSV file created after a call of EVExportUnderlying().  The data can be Sv values or values from, for example, the Echoview formula operator.
}
\details{
The default return option is a data.frame of ping header and sample information.  
If asList=TRUE a list with $header and $samples is returned,
If onlyHeader=TRUE only the headerinformation is returned.
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVExportUnderlying}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
