% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolRelativeDimension.R
\name{schoolRelativeDimension}
\alias{schoolRelativeDimension}
\title{Calculate the relative dimension of a school}
\usage{
schoolRelativeDimension(Sv, L, P, Th, theta)
}
\arguments{
\item{Sv}{Observed mean volume backscattering strength (Sv) of a school}

\item{L}{Observed (uncorrected) school length, m}

\item{P}{Mean depth of school}

\item{Th}{Minimum Sv data threshold used to detect schools using the SHAPES algorithm}

\item{theta}{Transducer 3 dB beam width, degrees}
}
\value{
Relative school dimension (Nb)
}
\description{
This function calculates the relative school dimension of Diner (2001)
}
\references{
Diner, N. (2001). Correction on school geometry and density: approach based on acoustic image simulation. Aquatic Living Resources, 14(4), 211-222.
}
\seealso{
\link{EVSchoolsDetect} \link{EVIntegrationByRegionsExport}
}
