% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVStartQuitEchoview.R
\name{StartEchoview}
\alias{StartEchoview}
\title{Start Echoview over COM}
\usage{
StartEchoview()
}
\value{
an Echoview Application object
}
\description{
This function opens an Echoview instance through COM scripting
}
\examples{
EVAppObj <- StartEchoview()
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
