% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateDataPathsAndECS.R
\name{UpdateDataPathAndECS}
\alias{UpdateDataPathAndECS}
\title{Title}
\usage{
UpdateDataPathAndECS(
  file.list,
  filesetName = "Primary Fileset",
  UpdateECS = TRUE
)
}
\arguments{
\item{file.list}{list of files}

\item{filesetName}{name of fileset}

\item{UpdateECS}{Logical stating whether or not to update the ECS file}
}
\value{
nothing
}
\description{
Title
}
