% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRevl2file.R
\name{writeRevl2file}
\alias{writeRevl2file}
\title{Write an Echoview line (.evl) file}
\usage{
writeRevl2file(evlObj, pathFn)
}
\arguments{
\item{evlObj}{An R object of class 'evl' resulting from a call of readevl2R()}

\item{pathFn}{path and filename of evl file to be create}
}
\value{
list object: [1]$success; [2] msg message log
}
\description{
Write an Echoview line (.evl) file
}
\seealso{
\code{\link{readevl2R}}
}
