% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVChangeVariableGrid}
\alias{EVChangeVariableGrid}
\title{Change the grid of an acoustic variable}
\usage{
EVChangeVariableGrid(EVFile, acousticVar, verticalType, horizontalType,
  verticalDistance = 0, horizontalDistance = 0, EVLine = NULL)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{acousticVar}{an EV acoustic variable object}

\item{verticalType}{0 = no grid, 1 = time (minutes), 2 = GPS distance (NMi), 3 = Vessel Log Distance (NMi), 4 = Pings, 5 = GPS distance (m), 6 = Vessel Log Distance (m).}

\item{horizontalType}{0 = no grid, 1 = depth grid, 2 = use reference line.}

\item{verticalDistance}{vertical grid line spacing. Not needed if verticalType = 0.}

\item{horizontalDistance}{horizontal grid line spacing. Not needed if horizontalType = 0.}

\item{EVLine}{an EV line object. Not needed if horizontalType = 0.}
}
\value{
a list object with two elements.  $dataRangeSettings: a vector of pre- and post-function call data range settings, and $msg: message for processing log.
}
\description{
This function sets the grid separation and depth reference line for an acoustic variable using COM scripting.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
varObj <- EVAcoVarNameFinder(EVFile, acoVarName = "38 seabed and surface excluded")$EVVar
EVLine <- EVFindLineByName(EVFile = EVFile, lineName = "Fixed depth 250 m")

#Change grid to 100m vertical distance and 10m depth grid relative to 100m line 
EVChangeVariableGrid(EVFile = EVFile, acousticVar = varObj, 
verticalType = 5, horizontalType = 2, verticalDistance = 100, horizontalDistance = 10, EVLine)
#remove horizontal and vertical grid
EVChangeVariableGrid(EVFile = EVFile, acousticVar = varObj, 
verticalType = 0, horizontalType = 0)
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}} \code{\link{EVAcoVarNameFinder}} \code{\link{EVFindLineByName}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
