% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVFishTracking.R
\name{EVFishTracking}
\alias{EVFishTracking}
\title{Detect Fish Tracks}
\usage{
EVFishTracking(
  EVFile,
  EVVar,
  FishTrackRegionClass = "Unclassified regions",
  export = NULL,
  DataDimensions = NULL,
  Alpha.MajorAxis = NULL,
  Alpha.MinorAxis = NULL,
  Alpha.Range = NULL,
  Beta.MajorAxis = NULL,
  Beta.MinorAxis = NULL,
  Beta.Range = NULL,
  ExclusionDistance.MajorAxis = NULL,
  ExclusionDistance.MinorAxis = NULL,
  ExclusionDistance.Range = NULL,
  MissedPingExpansion.MajorAxis = NULL,
  MissedPingExpansion.MinorAxis = NULL,
  MissedPingExpansion.Range = NULL,
  Weights.MajorAxis = NULL,
  Weights.MinorAxis = NULL,
  Weights.Range = NULL,
  Weights.TS = NULL,
  Weights.PingGap = NULL,
  MinimumTargets = NULL,
  MinimumPings = NULL,
  MaximumGap = NULL
)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{EVVar}{An Echoview Variable, accepts inputs as Character, list or Variable object (COMIDispatch)}

\item{LineName}{Character of the output name for the detected Line}

\item{algorithm}{numeric [0 - 2] Defines which bottom detection algorithm should be used: 0 for Delta Sv, 1 for Maximum Sv, 2 for Best bottom Candidate}

\item{StartDepth}{numeric [m] Minimum bottom detection depth}

\item{StopDepth}{numeric [m] maximum bottom detection detpth}

\item{MinSv}{numeric [dB] minimum detection Sv}

\item{UseBackstep}{Boolean [True or False]}

\item{BackstepRange}{numeric [m] Backstep range}

\item{DiscriminationLevel}{numeric [dB] Minimum discrimination threshold}

\item{MaxDropOuts}{numeric [samples] Maximum number of dropout samples before bottom detection fails}

\item{windowRadius}{numeric [samples] Search window size}

\item{PeakThreshold}{numeric [dB] Threshold for peak detection}

\item{MinPeakAssymmetry}{numeric}

\item{replaceOldBottom}{Boolean (TRUE or FALSE) If TRUE and a line with the same name as LineName already exists, the old line will be overwritten with the new one}
}
\description{
This function runs a bottom detection on an echogram
Bottom algorithms and settings are explained in the [Echoview Help File]('https://support.echoview.com/WebHelp/Reference/Algorithms/Line_picking_algorithm.htm')
}
\examples{
\dontrun{
#To be added - Needs Example data
#Starting Echoview
echoview = StartEchoview()
#Create a new EV File
EVFile <- EVCreateNew(EVAppObj=echoview, dataFiles = "rawfile.raw")$EVFile
Varname <- "Fileset 1: Sv pulse compressed wideband pings T1"
bottom <- EVBottomDetection(EVFile, EVVar=Varname, LineName="Bottom")
# Change the algorithm to Best bottom candidate
bottom <- EVBottomDetection(EVFile, EVVar=Varname, LineName="Bottom",algorithm=2)
}

}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
