% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVFreqSubset.R
\name{EVFreqSubset}
\alias{EVFreqSubset}
\title{Frequency subset from Wideband data.}
\usage{
EVFreqSubset(EVFile, acoVarName, frequency)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{acoVarName}{Acoustic variable name}

\item{frequency}{A Frequency contained within the Wideband Signal (see examples).}
}
\value{
a list object with two elements.  $success: ping subset successfully changed, and $msg: message for processing log.
}
\description{
This function subsets a single frequency from a wideband signal..
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
