% Generated by roxygen2 (4.0.1): do not edit by hand
\name{EVGetCalibrationFileName}
\alias{EVGetCalibrationFileName}
\title{Gets the calibration file name of a fileset}
\usage{
EVGetCalibrationFileName(EVFile, filesetName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{filesetName}{a string containing the name of the Echoview fileset}
}
\value{
an object: returns the calibration file name
}
\description{
This function gets the calibration file name of a filesset using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVGetCalibrationFileName(EVFile = EVFile, filesetName = "038-120-200")
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}

