% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVImportLine.R
\name{EVImportLine}
\alias{EVImportLine}
\title{Import an Echoview line (EVL format) file}
\usage{
EVImportLine(EVFile, pathAndFn = NULL, lineName = NULL,
  overwriteExistingLine = TRUE)
}
\arguments{
\item{EVFile}{("COMIDispatch) An Echoview file COM object}

\item{pathAndFn}{= NULL (character) path and file name of an EVL file}

\item{lineName}{= NULL (character) desired line name}

\item{overwriteExistingLine}{= TRUE (boolean) overwrite an existing line named lineName}
}
\value{
list $lineObj=COM object for the imported line (class COMIDispatch), msg= vector of messages)
}
\description{
This function imports an Echoview line (.EVL) format line into Echoview and can rename the imported line or replace a line.
}
\note{
Overwriting an existing line only works with editable lines.
}
