% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVImportLine.R
\name{EVImportLine}
\alias{EVImportLine}
\title{Imports an Echoview Line file
This function imports an Echoview line file (.evl) using COM scripting}
\usage{
EVImportLine(EVFile, pathAndFn = NULL, lineName = NULL)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{pathAndFn}{string path and filename to .evl file.}

\item{lineName}{=NULL optional line name for imported line.}
}
\value{
a list object with two elements- [[1]] imported EV line COM object, and [[2]] function message for log
}
\description{
Imports an Echoview Line file
This function imports an Echoview line file (.evl) using COM scripting
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj, 'c:\\\\Users\\\\martin_cox\\\\Desktop\\\\KOASLine.EV')$EVFile
#testing...
 EVImportLine(EVFile)
 test two...
EVImportLine(EVFile,pathAndFn='c:\\\\Users\\\\martin_cox\\\\Desktop\\\\lineKAOS.evl')

test three.. same line names (to be run after test2)
EVImportLine(EVFile,pathAndFn='c:\\\\Users\\\\martin_cox\\\\Desktop\\\\lineKAOS.evl',
lineName='Line6' )

test four.. rename line with unique name
lineInfo=EVImportLine(EVFile,pathAndFn='c:\\\\Users\\\\martin_cox\\\\Desktop\\\\lineKAOS.evl',
lineName='seabed4' )
lineInfo  
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVCreateEditableLine}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
