% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVImportRegionDef.R
\name{EVImportRegionDef}
\alias{EVImportRegionDef}
\title{Import an Echoview region definitions file (.evr)}
\usage{
EVImportRegionDef(EVFile, evrFile)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{evrFile}{An Echoview region definitions file (.evr) path and name}
}
\value{
If successful, A list object with three elements. $regionCOMObjs a list comprising of the COM object of each imported region;
$regionMeta metadata of the regions imported: the region ID, region name and region class, and $msg message for processing log
}
\description{
This function imports a region definitions file (.evr) using COM scripting
}
\examples{
\dontrun{EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj,'~~/KAOS/KAOStemplate.EV')$EVFile
EVImportRegionDef(EVFile = EVFile, evrFile = '~~/KAOS/off transect regions/20030114_1200000000.evr')
}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVOpenFile}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
