% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EchoviewR.R
\name{EVLineExportFromAcousticVarEVL}
\alias{EVLineExportFromAcousticVarEVL}
\title{Export an Echoview line file for an existing line associated with an Acoustic Variable.}
\usage{
EVLineExportFromAcousticVarEVL(EVFile, acoVar, lineNameToExport,
  pathAndFileName)
}
\arguments{
\item{EVFile}{An Echoview file COM object}

\item{acoVar}{Acoustic variable name that the line to be exported is associated with.}

\item{lineNameToExport}{a string containing the name of the line to export}

\item{pathAndFileName}{Path and filename of evl file.}
}
\description{
This function exports an EV line definition file (.evl) for a line associated with an acoustic variable.
}
\examples{
\dontrun{
EVAppObj <- COMCreate('EchoviewCom.EvApplication')
EVFile <- EVOpenFile(EVAppObj, '~\\\\example1.EV')$EVFile

EVLineExportFromAcousticVarEVL(EVFile=EVFile,
   acoVar='120 seabed and surface excluded',
   lineNameToExport='Fixed depth 250 m',
   pathAndFileName=  '~~KAOS/test2.evl')
}

}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\seealso{
\code{\link{EVFindLineByName}} \code{\link{EVCreateEditableLine}}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
