% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JollyAndHampton1990.R
\name{jhF}
\alias{jhF}
\title{The Jolly and Hampton (1990) estimator}
\usage{
jhF(transectLength, transectMeanDen, transectName = NULL, svyName = NULL,
  area = NULL)
}
\arguments{
\item{transectLength}{vector of parameter lengths}

\item{transectMeanDen}{transect weighted mean densities. See details.}

\item{transectName}{=NULL Optional character vector of transect names.}

\item{svyName}{=NULL Optional character variable of survey names.  Must be used if the \link{jhMultipleStrataF} function is used to estimate mean and variance across multiple strata.}

\item{area}{=NULL Optional character variable of the survey or stratum name.  Must be used if the \link{jhMultipleStrataF} function is used to estimate mean and variance across multiple strata.}
}
\value{
two element list.
}
\description{
The Jolly and Hampton (1990) estimator is often used to calculate mean pelagic animal densities and associated variances.
}
\details{
The \code{transectMeanDen} In the case of acoustic data, each element (areal density value) is multiplied by the integration interval (along transect) length.  The distance multiplication must take place outside of this function.
}
\examples{
\dontrun{
tLengths=c(77.67,78.28,78.91,80.22,78.07,73.60,78.53,78.39,78.52,79.48)
denV=c(158.03,159.50,96.41,133.86,246.60,243.95,122.65,276.92,59.83,89.90)
svyA=80*100*1e6
jhF(transectLength=tLengths,
transectMeanDen=denV,
transectName=NULL,svyName='test',area=svyA)
}
}
\references{
Jolly, G. M., and Hampton, I. 1990. A stratified random transect design for acoustic surveys of fish stocks. Canadian Journal of Fisheries and Aquatic Sciences, 47: 1282-1291.
}
\seealso{
\link{jhMultipleStrataF}
}
