% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVStartQuitEchoview.R
\name{QuitEchoview}
\alias{QuitEchoview}
\title{Quick Echoview over COM}
\usage{
QuitEchoview(EVAppObj)
}
\arguments{
\item{EVAppObj}{Echoview Application Object}
}
\value{
an Echoview Application object
}
\description{
This function closes an Echoview instance through COM scripting
}
\examples{
\dontrun{
EVAppObj=StartEchoview()
QuitEchoview(EVAppObj)

}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
