% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EVStartEchoview.R
\name{StartEchoview}
\alias{StartEchoview}
\title{Start an Echoview instance through COM scripting}
\usage{
StartEchoview()
}
\value{
an object of class COMDispatch
}
\description{
This function opens an Echoview instance
}
\examples{
\dontrun{
EVAppObj=StartEchoview()

}
}
\references{
\url{http://support.echoview.com/WebHelp/Echoview.htm/}
}
\keyword{COM}
\keyword{Echoview}
\keyword{scripting}
